Removed the need for a template for image generation. Just uses a default VM, which greatly simplifies the process.

+++ create_disk_image.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
 group_name="gtDiskCreator$(date +%Y%m%d%H%M%S)"
+vm_name="dcreator"
 template_file="image_generator.json"
 template_parameters="image_generator.parameters.json"
 location="eastus"
-
+docker_user=$(whoami)
+docker_password=""
+docker_image="gadgetron/ubuntu_1604_no_cuda"
 ME=$(basename "$0")
 
 print_usage ()
@@ -15,8 +18,9 @@ print_usage ()
     printf "Available options\n\n"
     printf "  -h | --help                          : Print help text\n"
     printf "  -g | --group <GROUP NAME>            : Name of ResourceGroup (default: $group_name)\n"
-    printf "  -t | --template <TEMPLATE FILE>      : Template file name (default: $template_file)\n"
-    printf "  -p | --parameters <PARAMETERS FILE>  : Template parameters file name (default: $template_parameters)\n"
+    printf "  -u | --docker-user <USERNAME>        : Docker username (default: $docker_user)\n"
+    printf "  -p | --docker-password <PASSWORD>    : Password for docker user\n"
+    printf "  -i | --docker-image <IMAGE>          : Docker image name (default: ${docker_image}\n"
     printf "  -l | --location <LOCATION>           : Location (default: $location)\n"
     printf "\n\n"
 }
@@ -30,11 +34,14 @@ while [[ $1 =~ ^- ]]; do
         -g  | --group )                shift
                                        group_name=$1
                                        ;;
-        -t  | --template )             shift
-                                       template_file=$1
+        -u  | --docker-user )          shift
+                                       docker_user=$1
                                        ;;
-        -p  | --parameters )           shift
-                                       template_parameters=$1
+        -p  | --docker-password )      shift
+                                       docker_password=$1
+                                       ;;
+        -i  | --docker-image )         shift
+                                       docker_image=$1
                                        ;;
         -l  | --location )             shift
                                        location=$1
@@ -46,23 +53,24 @@ while [[ $1 =~ ^- ]]; do
     shift
 done
 
+if [[ -z $docker_password ]]; then
+    echo "You must specify a Docker password"
+    exit 1
+fi
 
-time az group create --name ${group_name} --location ${location}
-time az group deployment create --resource-group ${group_name} --template-file ${template_file} --parameters @${template_parameters}
+az group create --name ${group_name} --location ${location}
+az vm create -g ${group_name} -n ${vm_name} --image UbuntuLTS --generate-ssh-keys
 
-while [ $(az vm show -g ${group_name} -n gtDiskCreator | jq -r .provisioningState) != "Succeeded" ]; do
-    echo "Waiting for VM to deploy"
-done
-
-sleep 20
+az vm extension set --publisher Microsoft.Azure.Extensions --name CustomScript --version 2.0 --settings "{\"fileUris\": [ \"https://raw.githubusercontent.com/hansenms/azure_templates/master/setup_disk_creator.sh\" ], \"commandToExecute\": \"bash ./setup_disk_creator.sh $docker_user $docker_password $docker_image\" }" --resource-group ${group_name} --vm-name ${vm_name}
 
 #Remove extension so that we can add a new one later. 
-extension_id=$(az vm extension list -g $group_name --vm-name gtDiskCreator | jq -r .[0].id)
+extension_id=$(az vm extension list -g ${group_name} --vm-name ${vm_name} | jq -r .[0].id)
 az vm extension delete --ids $extension_id 
+ipaddr=$(az vm list-ip-addresses -g ${group_name} -n ${vm_name} | jq -r .[0].virtualMachine.network.publicIpAddresses[0].ipAddress)
 
 #Deprovision
-command="sudo waagent -force -deprovision"
-ssh -o StrictHostKeyChecking=no gadgetron@${group_name}vm.${location}.cloudapp.azure.com $command
+command="sudo waagent -force -deprovision+user"
+ssh -o StrictHostKeyChecking=no $ipaddr $command
 
-sh ./create_image_from_vm.sh ${group_name}  gtDiskCreator
+sh ./create_image_from_vm.sh ${group_name} ${vm_name}
